componentize custom react table
