+++ airflow/ui/src/views/Pipelines/Row.tsx
@@ -17,6 +17,8 @@
  * under the License.
  */
 
+// Components to customize cell elements in the PipelinesTable
+
 import React from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 import {
